OrderViewSet filters added. Order, Tip model update.

+++ payments/models.py
@@ -5,4 +5,4 @@ from orders.models import Order
 
 class Tip(models.Model):
     order = models.OneToOneField(to=Order, on_delete=models.CASCADE, related_name="tip")
-    amount = models.PositiveIntegerField(blank=True)
+    amount = models.DecimalField(max_digits=10, decimal_places=2, blank=True)
